requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-dotenv
openrouteservice
folium
